fixed warnings: 10 valves don't fit in 1 byte

+++ invictus2/obc/include/filling_sm.h
@@ -136,7 +136,7 @@ struct filling_sm_object {
     union filling_data data;
 
     // TODO: Make this into an union
-    uint8_t valve_states;
+    uint16_t valve_states;
 
     struct filling_sm_config *config;
 };
